Got rid of most of the warnings compiler printed with -Wall
git-svn-id: http://merlin.fit.vutbr.cz/svn/STK/trunk@291 1d61b05a-ac01-0410-9a19-bfe7c875843c

+++ src/STKLib/Models.cc
@@ -2042,8 +2042,8 @@ namespace STK
   Print(std::ostream& rOstr)
   {
     FLOAT* p_vec = mpMatrixO;
-    int    i;
-    int    j;
+    size_t i;
+    size_t j;
 
     // go over all rows
     for (i=0; i < mNStates; i++) {
@@ -3876,7 +3876,7 @@ namespace STK
     mAllMixuresUpdatableFromStatAccums = true;
     
     if(mCmllrStats) {
-      for (int i=0; i < mNumberOfXformsToUpdate; i++) {
+      for (size_t i=0; i < mNumberOfXformsToUpdate; i++) {
         int size = mpXformToUpdate[i].mpXform->mInSize;
         mpXformToUpdate[i].mpXform->mpCmllrStats = (FLOAT *) calloc(sizeof(FLOAT), ((size+size*(size+1)/2) * (size-1) + 1));
       }
